Conflicts:
	ami/root/etc/init.d/spark

+++ lib/modules/cluster-destroy.sh
@@ -8,7 +8,12 @@ source $APP_ROOT/lib/common.sh
 MODULE_NAME="destroy"
 MODULE_DESCRIPTION="Destroy Cluster"
 
-echo "Deleting cluster $CFN_CLUSTER_NAME"
-aws cloudformation delete-stack --stack-name $CFN_CLUSTER_NAME >/dev/null
+echo "Destroy cluster $CFN_CLUSTER_NAME"
 
-aws_cfn_wait_for_quit_status "DELETE_COMPLETE" "ROLLBACK_COMPLETE"
+read -p "Are you sure you want to destroy cluster $CFN_CLUSTER_NAME? [yes,NO]: "
+if [ "x$REPLY" = "xyes" ]
+then
+    aws cloudformation delete-stack --stack-name $CFN_CLUSTER_NAME >/dev/null
+    aws_cfn_wait_for_quit_status "DELETE_COMPLETE" "ROLLBACK_COMPLETE"
+fi
+
